Structured the code closer to how we discussed
There is a fair amount of code in here to just get stuff working.
Archers are currently just upside down swordsman.
I've yet to implement JSON, so a lot of the information is hardcoded.
We should flesh out the rest of the class structures, since I had to
improvise on a lot of stuff, and re-wrote a bunch of stuff due to
unforseen issues.